Chore: abstract input validators from `bookFormHandler.js` to `/input-handlers/`
